[Rename]: `data_store` to `store`
