+++ src/components/navigation/Navigation.tsx
@@ -7,8 +7,8 @@
 import { NavLink } from "react-router-dom";
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { Theme, setTheme } from "../../data_store/slice/theme/ThemeSlice";
-import { RootState } from "../../data_store/Store";
+import { Theme, setTheme } from "../../store/slice/theme/ThemeSlice";
+import { RootState } from "../../store/Store";
 import NavigationToggleButton from "./NavigationToggleButton";
 import {
   ProfilePlaceholder,
