+++ src/provider/alert_toast_provider/AlertToastProvider.tsx
@@ -5,7 +5,7 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 import { useSelector } from "react-redux";
-import { RootState } from "../../data_store/Store";
+import { RootState } from "../../store/Store";
 import {
   AlertComponentData,
   AlertType,
