+++ src/screens/private/main_root_screen/MainRootScreen.tsx
@@ -6,15 +6,15 @@
 
 import { Outlet } from "react-router-dom";
 import { batch, useDispatch, useSelector } from "react-redux";
-import { onSignOut } from "../../../data_store/slice/auth/AuthSlice";
+import { onSignOut } from "../../../store/slice/auth/AuthSlice";
 import { signOut } from "../../../service/supabase/supa_auth/AuthApi";
 import { useCallback, useEffect, useState } from "react";
 import {
   removeProfile,
   syncForTheFirstTime,
   updateProfile,
-} from "../../../data_store/slice/profile/ProfileSlice";
-import { RootState } from "../../../data_store/Store";
+} from "../../../store/slice/profile/ProfileSlice";
+import { RootState } from "../../../store/Store";
 import { NavigationComponent, SidebarComponent } from "components";
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
