+++ src/screens/private/screens/profile/profile_preview/ProfilePreviewScreen.tsx
@@ -5,7 +5,7 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 import { useSelector } from "react-redux";
-import { RootState } from "../../../../../data_store/Store";
+import { RootState } from "../../../../../store/Store";
 import {
   SocialLink,
   SocialType,
