+++ src/screens/private/screens/projects/ProjectScreen.tsx
@@ -8,10 +8,10 @@ import { NavLink, Outlet, useLocation } from "react-router-dom";
 import { TooltipComponent } from "components";
 import getSubNavTitle from "../../../../util/nav/NavTitle";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../../../../data_store/Store";
+import { RootState } from "../../../../store/Store";
 import { useEffect } from "react";
 import { getAllProjects } from "../../../../service/supabase/supastore/projects/ProjectsStoreApi";
-import { addProjects } from "../../../../data_store/slice/projects/ProjectsSlice";
+import { addProjects } from "../../../../store/slice/projects/ProjectsSlice";
 import { AddIcon, DeleteIcon, EditIcon, PreviewIcon } from "assets";
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
