+++ src/screens/private/screens/projects/project_delete/ProjectDeleteScreen.tsx
@@ -6,7 +6,7 @@
 
 import { useSelector } from "react-redux";
 
-import { RootState } from "../../../../../data_store/Store";
+import { RootState } from "../../../../../store/Store";
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { ProjectCard, getProjectCardData } from "components";
