+++ src/screens/private/screens/projects/project_preview/ProjectPreviewComponent.tsx
@@ -7,7 +7,7 @@
 
 import { useDispatch } from "react-redux";
 import { ImagePreviewModalButton } from "components";
-import { setModalImage } from "../../../../../data_store/slice/image_modal/ImageModalSlice";
+import { setModalImage } from "../../../../../store/slice/image_modal/ImageModalSlice";
 import { ProjectData } from "../../../../../service/supabase/supastore/projects/ProjectsCollection";
 import { WebsiteIcon, GithubIcon, HostingerIcon } from "assets";
