+++ src/screens/public/login_screen/LoginScreen.tsx
@@ -10,7 +10,7 @@ import { ErrorText, InputText, InputType, LandingIntro } from "components";
 
 import { LoginCred } from "../../../service/supabase/supa_auth/actions/AuthSignIn";
 import { signIn } from "../../../service/supabase/supa_auth/AuthApi";
-import { AuthData, onLogin } from "../../../data_store/slice/auth/AuthSlice";
+import { AuthData, onLogin } from "../../../store/slice/auth/AuthSlice";
 import { useDispatch } from "react-redux";
 import { validatePassword } from "../../../util/input/Input";
