+++ src/screens/public/signup_screen/SignupScreen.tsx
@@ -9,7 +9,7 @@ import { NavLink, useNavigate } from "react-router-dom";
 import { ErrorText, InputText, LandingIntro } from "components";
 import { SignUpCred } from "../../../service/supabase/supa_auth/actions/AuthSignUp";
 import { signUp } from "../../../service/supabase/supa_auth/AuthApi";
-import { showAlertWithTimeout } from "../../../data_store/slice/toast_alert/ToastAlertSlice";
+import { showAlertWithTimeout } from "../../../store/slice/toast_alert/ToastAlertSlice";
 import { useDispatch } from "react-redux";
 import { getSuccessAlertData } from "../../../provider/alert_toast_provider/AlertToastProvider";
