+++ src/service/local_storage/auth/AuthStorageApi.ts
@@ -4,7 +4,7 @@
  */
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-import { AuthData } from "../../../data_store/slice/auth/AuthSlice";
+import { AuthData } from "../../../store/slice/auth/AuthSlice";
 import { AuthTokenKey } from "../token_constants/StorageConstant";
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
