+++ src/service/local_storage/theme/ThemeStorageApi.ts
@@ -4,7 +4,7 @@
  */
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-import { Theme } from "../../../data_store/slice/theme/ThemeSlice";
+import { Theme } from "../../../store/slice/theme/ThemeSlice";
 import { themeKey } from "../token_constants/StorageConstant";
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
