+++ src/service/supabase/supa_auth/AuthApi.ts
@@ -3,7 +3,7 @@
  * Last updated: 04 Feb, 2024
  */
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-import { AuthData } from "../../../data_store/slice/auth/AuthSlice";
+import { AuthData } from "../../../store/slice/auth/AuthSlice";
 import { AuthTokenKey } from "../../local_storage/token_constants/StorageConstant";
 import { firebaseForgotPassword } from "./actions/AuthForgotPassword";
 import { LoginCred, firebaseSingIn } from "./actions/AuthSignIn";
