+++ src/service/supabase/supa_auth/actions/AuthSignIn.ts
@@ -5,7 +5,7 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 import { SupabaseAuth } from "../../supabase_main/Supabase";
-import { AuthData } from "../../../../data_store/slice/auth/AuthSlice";
+import { AuthData } from "../../../../store/slice/auth/AuthSlice";
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 /**
