+++ src/service/supabase/supastore/user_profile/actions/UserDataRead.ts
@@ -4,7 +4,7 @@
  */
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-import { UserData } from "../../../../../data_store/slice/auth/AuthSlice";
+import { UserData } from "../../../../../store/slice/auth/AuthSlice";
 import { SupabaseUsersDB } from "../../../supabase_main/Supabase";
 import { SocialLink, SocialType, UserProfileData } from "../UserCollection";
