+++ src/store/slice/image_modal/ImageModalSlice.tsx
