+++ src/store/slice/image_modal/index.ts
